fixup "sculpt: add File_handler utility" (32-bit build)

+++ repos/gems/src/app/sculpt_manager/file_handler.h
@@ -107,7 +107,7 @@ struct Genode::File_handler
 
 			for (unsigned i = 0; i < MAX_ATTEMPTS; i++) {
 
-				try { _file_size = _dir.file_size(_path); }
+				try { _file_size = size_t(_dir.file_size(_path)); }
 				catch (...) {
 					disappeared = true;
 					_file_size = 0;
